README.md
What I have learned from github is that this is a platform for version control of your code along with great opportunities to collab with my fellow tablemates. In addition, I learned the basics of commiting my code, so I can record changes to files in my branch and how to post repositories to show my contributions to code.